distmaker - Reduce noise in dists/*.sh

+++ distmaker/dists/drupal6_php5.sh
@@ -1,25 +1,18 @@
 #!/bin/bash
 set -ex
 
-# This script assumes
-# that DAOs are generated
-# and all the necessary conversions had place!
-
 P=`dirname $0`
 CFFILE=$P/../distmaker.conf
-
 if [ ! -f $CFFILE ] ; then
 	echo "NO DISTMAKER.CONF FILE!"
 	exit 1
 else
 	. $CFFILE
 fi
-
 . "$P/common.sh"
 
 SRC=$DM_SOURCEDIR
 TRG=$DM_TMPDIR/civicrm
-
 
 # checkout the right code revisions
 pushd "$DM_SOURCEDIR/drupal"
@@ -28,9 +21,7 @@ git checkout "$DM_REF_DRUPAL6"
 popd
 
 # make sure and clean up before
-if [ -d $TRG ] ; then
-	rm -rf $TRG/*
-fi
+[ -d $TRG ] && rm -rf $TRG/*
 
 # copy all the stuff
 dm_install_core "$SRC" "$TRG"
