+++ distmaker/dists/l10n.sh
@@ -1,29 +1,21 @@
 #!/bin/bash
 set -ex
 
-# This script assumes
-# that DAOs are generated
-# and all the necessary conversions had place!
-
 P=`dirname $0`
 CFFILE=$P/../distmaker.conf
-
 if [ ! -f $CFFILE ] ; then
 	echo "NO DISTMAKER.CONF FILE!"
 	exit 1
 else
 	. $CFFILE
 fi
-
 . "$P/common.sh"
 
 SRC=$DM_SOURCEDIR
 TRG=$DM_TMPDIR/civicrm
 
 # make sure and clean up before
-if [ -d $TRG ] ; then
-	rm -rf $TRG/*
-fi
+[ -d $TRG ] && rm -rf $TRG/*
 
 # copy all the stuff
 dm_install_l10n "$SRC/l10n" "$TRG/l10n"
